create SimpleTable with data and headers props
